refactor(blorktools): zip handler consolidation
